requirements.txt
pytesseract
pillow
gunicorn
